+++ build.gradle
@@ -21,8 +21,8 @@ compileJava {
     sourceCompatibility = '1.8'
     targetCompatibility = '1.8'
 }
-version = "1.0.1"
-group= "dev.meyi"
+version = "1.1.0"
+group= "dev.meyi.bn"
 archivesBaseName = "BazaarNotifier"
 
 minecraft {
