+++ src/main/java/dev/meyi/bn/EventHandler.java
@@ -1,4 +1,4 @@
-package dev.meyi;
+package dev.meyi.bn;
 
 import net.minecraft.client.Minecraft;
 import net.minecraft.util.ChatComponentText;
